Changing student id numbers to be generated by the id generator

+++ src/com/endicott/edu/datalayer/IdNumberGenDao.java
@@ -56,7 +56,7 @@ public class IdNumberGenDao {
      * @param runId college instance ID
      * @return unique id
      */
-    public static long getID(String runId){
+    public static int getID(String runId){
        // logger.info("Fetching ID...");
         Gson gson = new Gson();
         ID id = new ID(-1);
